Removed applicants store
