+++ src/components/profiles/profile-card.vue
@@ -89,7 +89,7 @@ export default {
     ...mapActions('profiles', ['getVoiceToken']),
     ...mapActions('profiles', ['getPublicProfile']),
     ...mapActions('treasury', ['getSupply']),
-    ...mapActions('applicants', ['enroll']),
+    ...mapActions('accounts', ['enrollMember']),
 
     // How do we optimize this repeated profile requests?
     async getProfileDataFromContract () {
@@ -131,7 +131,7 @@ export default {
       event.stopPropagation()
       this.submittingEnroll = true
       try {
-        const res = await this.enroll({
+        const res = await this.enrollMember({
           applicant: this.username,
           content: 'DAO Enroll member'
         })
