+++ src/store/accounts/actions.js
@@ -216,6 +216,23 @@ export const applyMember = async function ({ state, rootState, commit }, { conte
   return result
 }
 
+export const enrollMember = async function ({ commit, rootState }, { applicant, content }) {
+  const selectedDao = this.getters['dao/selectedDao']
+
+  const actions = [{
+    account: this.$config.contracts.dao,
+    name: 'enroll',
+    data: {
+      enroller: rootState.accounts.account,
+      applicant,
+      dao_id: selectedDao.docId,
+      content
+    }
+  }]
+  const result = await this.$api.signTransaction(actions)
+  return result
+}
+
 export const checkMembership = async function ({ commit, state, dispatch }) {
   const selectedDao = this.getters['dao/selectedDao']
