--- src/store/applicants/mutations.js
@@ -1,1 +1,0 @@
-
