--- src/store/applicants/state.js
@@ -1,2 +1,0 @@
-export default () => ({
-})
